make logger its own thing.
