+++ src_client/js/lib/game.js
@@ -1,5 +1,4 @@
-var log4js = require('log4js');
-var logger = log4js.getLogger('Game');
+var logger = require('./logger.js');
 
 var THREE = require('../vendor/three.min.js');
 var rectShape = require('./game/rect_shape.js');
