+++ src_client/js/main.js
@@ -1,19 +1,4 @@
-var log4js = require('log4js');
-var logger = log4js.getLogger('Main');
-
-log4js.configure({
-  appenders: [
-    {
-      type: 'console',
-      layout: {
-        type: 'pattern',
-        pattern: "[%5.5p] - [%c] - %m%n"
-      }
-    }
-  ]
-});
-
-// logger.setLevel('OFF');
+var logger = require('./lib/logger.js');
 
 logger.info("Hi from main.js");
